+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.3.4 - Unreleased
  *  Fix bugs with symfony dependencies >= 4.2
  *  Add behat step to populate a repository during tests
  *  Add behat step to verify if multiple records exists at once
+ *  Fix bug with fixtures names containing special chars like `/\-_`
  
 v1.3.3 - 2019-04-24
 ----